Add redirector script from the old dartdocs.org path schema
The old schema was something like
index.html#package/library.Class@id_method. So, we parse it and try to
redirect to the new dartdoc's path. For that, we first make a HEAD
request to check if that location actually exists, and if yes - change
location.href to that new location

It doesn't always work, because there are some changes between the old
schema and the new schema, so sometimes the redirect would lead to 404.
Known cases:

Operators. Old dartdocs.org doesn't make a distinction between
an operator and e.g. a method, but the new dartdoc does. So, there is
no way to know if this is an operator or not, so the redirect won't
work.

Inherited methods. If class B extends A, and B has inherited methods
from A, then dartdocs.org will show them as #package/lib.B@id_foo, and
dartdoc will consider it as lib/A/foo.html, so the redirect won't work
either.

There may be other cases, but most of the time it should work fine.